feat: maintenance mode for web
