+++ templates/error/503.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Maintenance | SimpleLogin</title>
+    <link rel="icon" href="/static/favicon.ico" type="image/x-icon" />
+    <link rel="stylesheet" href="/static/node_modules/font-awesome/css/font-awesome.css" />
+    <link rel="stylesheet" href="/static/assets/css/dashboard.css" />
+  </head>
+  <body class="antialiased">
+    <div class="page-content">
+      <div class="container text-center">
+        <div class="display-3 text-muted mb-5">
+          <i class="fa fa-wrench"></i>
+          503
+        </div>
+        <h3 class="h3 mb-4">
+          SimpleLogin is currently undergoing scheduled maintenance.
+          <br />
+          Please try again in a few moments.
+        </h3>
+        <a class="btn btn-primary" href="javascript:location.reload()">
+          <i class="fa fa-refresh mr-2"></i>Refresh Page
+        </a>
+      </div>
+    </div>
+  </body>
+</html>
